Fix debug color with vertex colors
