+++ exts/cesium.omniverse/mdl/cesium.mdl
@@ -15,10 +15,29 @@ float4 alpha_blend(float4 src, float4 dst) {
     return src * float4(src.w, src.w, src.w, 1.0) + dst * (1.0 - src.w);
 }
 
+float4 compute_base_color(
+    gltf_texture_lookup_value imagery_layers_texture,
+    color debug_color,
+    gltf_texture_lookup_value base_color_texture,
+    color base_color_factor,
+    float base_alpha) {
+
+    auto debug_color_float3 = float3(debug_color);
+    auto base_color_factor_float3 = float3(base_color_factor);
+
+    auto base_color = base_color_texture.valid ? base_color_texture.value : float4(1.0);
+    base_color *= float4(base_color_factor_float3.x, base_color_factor_float3.y, base_color_factor_float3.z, base_alpha);
+    base_color = alpha_blend(imagery_layers_texture.value, base_color);
+    base_color *= float4(debug_color_float3.x, debug_color_float3.y, debug_color_float3.z, 1.0);
+
+    return base_color;
+}
+
 export gltf_texture_lookup_value cesium_texture_lookup(*) [[ anno::hidden() ]] = gltf_texture_lookup();
 
 export material cesium_material(
     gltf_texture_lookup_value imagery_layers_texture = gltf_texture_lookup_value(true, float4(0.0)),
+    uniform color debug_color = color(1.0),
     // gltf_material inputs below
     gltf_texture_lookup_value base_color_texture = gltf_texture_lookup_value(),
     uniform color base_color_factor = color(1.0),
@@ -29,20 +48,15 @@ export material cesium_material(
     uniform float base_alpha = 1.0,
     uniform float alpha_cutoff  = 0.5
 ) [[ anno::hidden() ]] = let {
-    auto base_color_texture_value = base_color_texture.valid ? base_color_texture.value : float4(1.0);
-    auto final_base_color_texture_value = alpha_blend(imagery_layers_texture.value, base_color_texture_value);
-    auto final_base_color_texture = gltf_texture_lookup_value(true, final_base_color_texture_value);
-    auto final_base_color_factor = base_color_factor;
-
+    auto base_color = compute_base_color(imagery_layers_texture, debug_color, base_color_texture, base_color_factor, base_alpha);
     material base = gltf_material(
-        base_color_factor: final_base_color_factor,
-        base_color_texture: final_base_color_texture,
+        base_color_texture: gltf_texture_lookup_value(true, base_color),
         metallic_factor: metallic_factor,
         roughness_factor: roughness_factor,
         emissive_factor: emissive_factor,
         alpha_mode: alpha_mode,
-        base_alpha: base_alpha,
-        alpha_cutoff: alpha_cutoff
+        alpha_cutoff: alpha_cutoff,
+        vertex_color_name: string("vertexColor")
     );
 
 } in material(
