+++ src/core/include/cesium/omniverse/FabricMaterialPool.h
@@ -16,6 +16,7 @@ class FabricMaterialPool final : public ObjectPool<FabricMaterial> {
         uint64_t initialCapacity,
         const pxr::TfToken& defaultTextureAssetPathToken,
         const pxr::TfToken& defaultTransparentTextureAssetPathToken,
+        bool debugRandomColors,
         long stageId);
 
     [[nodiscard]] const FabricMaterialDefinition& getMaterialDefinition() const;
@@ -29,6 +30,7 @@ class FabricMaterialPool final : public ObjectPool<FabricMaterial> {
     const FabricMaterialDefinition _materialDefinition;
     const pxr::TfToken _defaultTextureAssetPathToken;
     const pxr::TfToken _defaultTransparentTextureAssetPathToken;
+    const bool _debugRandomColors;
     const long _stageId;
 };
