+++ src/core/src/FabricGeometryPool.cpp
@@ -8,12 +8,10 @@ FabricGeometryPool::FabricGeometryPool(
     int64_t poolId,
     const FabricGeometryDefinition& geometryDefinition,
     uint64_t initialCapacity,
-    bool debugRandomColors,
     long stageId)
     : ObjectPool<FabricGeometry>()
     , _poolId(poolId)
     , _geometryDefinition(geometryDefinition)
-    , _debugRandomColors(debugRandomColors)
     , _stageId(stageId) {
     setCapacity(initialCapacity);
 }
@@ -25,7 +23,7 @@ const FabricGeometryDefinition& FabricGeometryPool::getGeometryDefinition() cons
 std::shared_ptr<FabricGeometry> FabricGeometryPool::createObject(uint64_t objectId) {
     const auto pathStr = fmt::format("/fabric_geometry_pool_{}_object_{}", _poolId, objectId);
     const auto path = omni::fabric::Path(pathStr.c_str());
-    return std::make_shared<FabricGeometry>(path, _geometryDefinition, _debugRandomColors, _stageId);
+    return std::make_shared<FabricGeometry>(path, _geometryDefinition, _stageId);
 }
 
 void FabricGeometryPool::setActive(std::shared_ptr<FabricGeometry> geometry, bool active) {
