+++ src/core/src/FabricMaterial.cpp
@@ -10,12 +10,16 @@
 #include "cesium/omniverse/Tokens.h"
 #include "cesium/omniverse/UsdUtil.h"
 
+#include <glm/gtc/random.hpp>
 #include <omni/fabric/FabricUSD.h>
 #include <spdlog/fmt/fmt.h>
 
 namespace cesium::omniverse {
 
 namespace {
+
+const auto DEFAULT_DEBUG_COLOR = pxr::GfVec3f(1.0f, 1.0f, 1.0f);
+
 uint64_t getImageryLayerCount(const FabricMaterialDefinition& materialDefinition) {
     uint64_t imageryLayerCount = materialDefinition.getImageryLayerCount();
 
@@ -31,6 +35,7 @@ uint64_t getImageryLayerCount(const FabricMaterialDefinition& materialDefinition
 
     return imageryLayerCount;
 }
+
 } // namespace
 
 FabricMaterial::FabricMaterial(
@@ -38,11 +43,13 @@ FabricMaterial::FabricMaterial(
     const FabricMaterialDefinition& materialDefinition,
     const pxr::TfToken& defaultTextureAssetPathToken,
     const pxr::TfToken& defaultTransparentTextureAssetPathToken,
+    bool debugRandomColors,
     long stageId)
     : _materialPath(path)
     , _materialDefinition(materialDefinition)
     , _defaultTextureAssetPathToken(defaultTextureAssetPathToken)
     , _defaultTransparentTextureAssetPathToken(defaultTransparentTextureAssetPathToken)
+    , _debugRandomColors(debugRandomColors)
     , _stageId(stageId) {
 
     if (stageDestroyed()) {
@@ -140,8 +147,6 @@ void FabricMaterial::createMaterial(const omni::fabric::Path& materialPath) {
 }
 
 void FabricMaterial::createShader(const omni::fabric::Path& shaderPath, const omni::fabric::Path& materialPath) {
-    const auto hasVertexColors = _materialDefinition.hasVertexColors();
-
     auto srw = UsdUtil::getFabricStageReaderWriter();
 
     srw.createPrim(shaderPath);
@@ -149,6 +154,7 @@ void FabricMaterial::createShader(const omni::fabric::Path& shaderPath, const om
     FabricAttributesBuilder attributes;
 
     // clang-format off
+    attributes.addAttribute(FabricTypes::inputs_debug_color, FabricTokens::inputs_debug_color);
     attributes.addAttribute(FabricTypes::inputs_alpha_cutoff, FabricTokens::inputs_alpha_cutoff);
     attributes.addAttribute(FabricTypes::inputs_alpha_mode, FabricTokens::inputs_alpha_mode);
     attributes.addAttribute(FabricTypes::inputs_base_alpha, FabricTokens::inputs_base_alpha);
@@ -167,10 +173,6 @@ void FabricMaterial::createShader(const omni::fabric::Path& shaderPath, const om
     attributes.addAttribute(FabricTypes::_cesium_tilesetId, FabricTokens::_cesium_tilesetId);
     // clang-format on
 
-    if (hasVertexColors) {
-        attributes.addAttribute(FabricTypes::inputs_vertex_color_name, FabricTokens::inputs_vertex_color_name);
-    }
-
     attributes.createAttributes(shaderPath);
 
     srw.setArrayAttributeSize(shaderPath, FabricTokens::_paramColorSpace, 0);
@@ -188,14 +190,6 @@ void FabricMaterial::createShader(const omni::fabric::Path& shaderPath, const om
     infoMdlSourceAssetFabric->assetPath = Context::instance().getCesiumMdlPathToken();
     infoMdlSourceAssetFabric->resolvedPath = pxr::TfToken();
     *infoMdlSourceAssetSubIdentifierFabric = FabricTokens::cesium_material;
-
-    if (hasVertexColors) {
-        const auto vertexColorPrimvarNameSize = pxr::UsdTokens->vertexColor.GetString().size();
-        srw.setArrayAttributeSize(shaderPath, FabricTokens::inputs_vertex_color_name, vertexColorPrimvarNameSize);
-        auto vertexColorNameFabric =
-            srw.getArrayAttributeWr<uint8_t>(shaderPath, FabricTokens::inputs_vertex_color_name);
-        memcpy(vertexColorNameFabric.data(), pxr::UsdTokens->vertexColor.GetText(), vertexColorPrimvarNameSize);
-    }
 
     // Connect the material terminals to the shader.
     srw.createConnection(
@@ -384,6 +378,7 @@ void FabricMaterial::clearImageryLayers() {
 void FabricMaterial::setShaderValues(const omni::fabric::Path& shaderPath, const MaterialInfo& materialInfo) {
     auto srw = UsdUtil::getFabricStageReaderWriter();
 
+    auto debugColorFabric = srw.getAttributeWr<pxr::GfVec3f>(shaderPath, FabricTokens::inputs_debug_color);
     auto alphaCutoffFabric = srw.getAttributeWr<float>(shaderPath, FabricTokens::inputs_alpha_cutoff);
     auto alphaModeFabric = srw.getAttributeWr<int>(shaderPath, FabricTokens::inputs_alpha_mode);
     auto baseAlphaFabric = srw.getAttributeWr<float>(shaderPath, FabricTokens::inputs_base_alpha);
@@ -399,6 +394,15 @@ void FabricMaterial::setShaderValues(const omni::fabric::Path& shaderPath, const
     *emissiveFactorFabric = UsdUtil::glmToUsdVector(glm::fvec3(materialInfo.emissiveFactor));
     *metallicFactorFabric = static_cast<float>(materialInfo.metallicFactor);
     *roughnessFactorFabric = static_cast<float>(materialInfo.roughnessFactor);
+
+    if (_debugRandomColors) {
+        const auto r = glm::linearRand(0.0f, 1.0f);
+        const auto g = glm::linearRand(0.0f, 1.0f);
+        const auto b = glm::linearRand(0.0f, 1.0f);
+        *debugColorFabric = pxr::GfVec3f(r, g, b);
+    } else {
+        *debugColorFabric = DEFAULT_DEBUG_COLOR;
+    }
 }
 
 void FabricMaterial::setTextureValues(
