+++ src/core/src/FabricMaterialPool.cpp
@@ -10,12 +10,14 @@ FabricMaterialPool::FabricMaterialPool(
     uint64_t initialCapacity,
     const pxr::TfToken& defaultTextureAssetPathToken,
     const pxr::TfToken& defaultTransparentTextureAssetPathToken,
+    bool debugRandomColors,
     long stageId)
     : ObjectPool<FabricMaterial>()
     , _poolId(poolId)
     , _materialDefinition(materialDefinition)
     , _defaultTextureAssetPathToken(defaultTextureAssetPathToken)
     , _defaultTransparentTextureAssetPathToken(defaultTransparentTextureAssetPathToken)
+    , _debugRandomColors(debugRandomColors)
     , _stageId(stageId) {
     setCapacity(initialCapacity);
 }
@@ -28,7 +30,12 @@ std::shared_ptr<FabricMaterial> FabricMaterialPool::createObject(uint64_t object
     const auto pathStr = fmt::format("/fabric_material_pool_{}_object_{}", _poolId, objectId);
     const auto path = omni::fabric::Path(pathStr.c_str());
     return std::make_shared<FabricMaterial>(
-        path, _materialDefinition, _defaultTextureAssetPathToken, _defaultTransparentTextureAssetPathToken, _stageId);
+        path,
+        _materialDefinition,
+        _defaultTextureAssetPathToken,
+        _defaultTransparentTextureAssetPathToken,
+        _debugRandomColors,
+        _stageId);
 }
 
 void FabricMaterialPool::setActive(std::shared_ptr<FabricMaterial> material, bool active) {
